Rename interactive-drive-prepare -> omnidreams-prepare
The script stages resources used by *both* demos (scene USDZs from
``nvidia/omni-dreams-scenes`` for interactive-drive + webrtc, and the
Cosmos-Reason1 text encoder for the world-model backend that both
demos exercise), so naming it ``interactive-drive-prepare`` implied a
single-demo scope it never had.

Move the module from ``omnidreams.interactive_drive.prepare`` up to
``omnidreams.prepare`` to match the new package-level scope, and pick
``omnidreams-prepare`` as the script name to follow the workspace's
``<feature>-<action>`` convention (``flashdreams-run``,
``flashvsr-grpc-server``, etc.) so the binary stays unambiguous if it
ever ends up on ``PATH`` outside ``uv run --package``.

Updates every README / docstring / SystemExit hint reference and the
auto-stage lazy import in ``interactive_drive.demo``. No behaviour
change beyond the rename + module move.
